+++ backend/tests/integration/tests/query_history/test_query_history.py
